refactor: 将矩阵相关的方法统一放置与 util/matrix 下
