+++ src/util/matrix.ts
@@ -1,3 +1,4 @@
+import { IElement } from '@antv/g-base';
 import { ext, vec2, vec3 } from '@antv/matrix-util';
 import { BBox, Point } from '../types';
 
@@ -60,4 +61,16 @@ export function applyMatrix2BBox(matrix: number[], bbox: BBox) {
     width: maxX - minX,
     height: maxY - minY,
   };
+}
+
+export function applyRotate(shape: IElement, rotate: number, x: number, y: number) {
+  if (rotate) {
+    const matrix = getMatrixByAngle({ x, y }, rotate, shape.getMatrix());
+    shape.setMatrix(matrix);
+  }
+}
+
+export function applyTranslate(shape: IElement, x: number, y: number) {
+  const translateMatrix = getMatrixByTranslate({ x, y });
+  shape.attr('matrix', translateMatrix);
 }
